19: Moving ProjectsPreview component to folder
