+++ src/components/ProjectsPreview/style.tsx
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const Heading = styled.h3`
+  margin-bottom: 2rem;
+`;
+
+export const ProjectsList = styled.ul`
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  align-items: center;
+  flex-wrap: wrap;
+  padding: 0;
+  list-style: none;
+`;
